# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Industry = (typeof Industry)[keyof typeof Industry];

export const Industry = {
  salon_spa: "salon_spa",
  real_estate: "real_estate",
  clinic_healthcare: "clinic_healthcare",
  restaurant_cafe: "restaurant_cafe",
  other_service: "other_service",
  other: "other",
} as const;

export type BudgetBand = (typeof BudgetBand)[keyof typeof BudgetBand];

export const BudgetBand = {
  under_1l: "under_1l",
  "1l_to_3l": "1l_to_3l",
  "3l_to_8l": "3l_to_8l",
  "8l_to_20l": "8l_to_20l",
  "20l_plus": "20l_plus",
  not_sure: "not_sure",
} as const;

export type Timeline = (typeof Timeline)[keyof typeof Timeline];

export const Timeline = {
  asap: "asap",
  within_1_month: "within_1_month",
  "1_to_3_months": "1_to_3_months",
  "3_to_6_months": "3_to_6_months",
  exploring: "exploring",
} as const;

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export type LeadInputIndustryDetails = { [key: string]: unknown };

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @nullable */
  businessName?: string | null;
  /** @minLength 6 */
  whatsappNumber: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 2 */
  city: string;
  industry: Industry;
  industryDetails?: LeadInputIndustryDetails;
  /** @minLength 10 */
  problem: string;
  /** @nullable */
  currentSolution?: string | null;
  /** @minLength 5 */
  goalIn3Months: string;
  budget: BudgetBand;
  timeline: Timeline;
}

export interface LeadUpdate {
  status?: LeadStatus;
  /** @nullable */
  notes?: string | null;
}

export type LeadIndustryDetails = { [key: string]: unknown };

export interface Lead {
  id: string;
  name: string;
  /** @nullable */
  businessName?: string | null;
  whatsappNumber: string;
  /** @nullable */
  email?: string | null;
  city: string;
  industry: Industry;
  industryDetails: LeadIndustryDetails;
  problem: string;
  /** @nullable */
  currentSolution?: string | null;
  goalIn3Months: string;
  budget: BudgetBand;
  timeline: Timeline;
  status: LeadStatus;
  /** @nullable */
  notes?: string | null;
  whatsappNotificationSent: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface SubscriberInput {
  email: string;
  /** @minLength 1 */
  source: string;
}

export interface Subscriber {
  id: string;
  email: string;
  source: string;
  createdAt: string;
}

export interface SiteStats {
  activeProjects: number;
  projectsShipped: number;
  leadsAllTime: number;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}