# lib/api-zod/src/generated/types/budgetBand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export type BudgetBand = (typeof BudgetBand)[keyof typeof BudgetBand];

export const BudgetBand = {
  under_1l: "under_1l",
  "1l_to_3l": "1l_to_3l",
  "3l_to_8l": "3l_to_8l",
  "8l_to_20l": "8l_to_20l",
  "20l_plus": "20l_plus",
  not_sure: "not_sure",
} as const;