# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetBand } from "./budgetBand";
import type { Industry } from "./industry";
import type { LeadInputIndustryDetails } from "./leadInputIndustryDetails";
import type { Timeline } from "./timeline";

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @nullable */
  businessName?: string | null;
  /** @minLength 6 */
  whatsappNumber: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 2 */
  city: string;
  industry: Industry;
  industryDetails?: LeadInputIndustryDetails;
  /** @minLength 10 */
  problem: string;
  /** @nullable */
  currentSolution?: string | null;
  /** @minLength 5 */
  goalIn3Months: string;
  budget: BudgetBand;
  timeline: Timeline;
}