# lib/api-zod/src/generated/types/leadStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;