# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface LeadUpdate {
  status?: LeadStatus;
  /** @nullable */
  notes?: string | null;
}