# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  activeProjects: number;
  projectsShipped: number;
  leadsAllTime: number;
}