# lib/api-zod/src/generated/types/subscriber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export interface Subscriber {
  id: string;
  email: string;
  source: string;
  createdAt: Date;
}