# lib/api-zod/src/generated/types/subscriberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriberInput {
  email: string;
  /** @minLength 1 */
  source: string;
}