# lib/api-zod/src/generated/types/timeline.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Everyday Digital Solutions
 * OpenAPI spec version: 0.1.0
 */

export type Timeline = (typeof Timeline)[keyof typeof Timeline];

export const Timeline = {
  asap: "asap",
  within_1_month: "within_1_month",
  "1_to_3_months": "1_to_3_months",
  "3_to_6_months": "3_to_6_months",
  exploring: "exploring",
} as const;